requirements.txt
fastapi
python-dotenv
httpx
uvicorn
python-multipart
pdfplumber
requests
playwright==1.41.0
apscheduler==3.10.4
pytz==2024.1
